Update flang compiler package to 20.1.7

+++ recipes/recipes/flang_emscripten-wasm32/activate.sh
@@ -3,9 +3,12 @@
 # # Using flang as a WASM cross-compiler
 # # https://github.com/serge-sans-paille/llvm-project/blob/feature/flang-wasm/README.wasm.md
 # # https://github.com/conda-forge/flang-feedstock/pull/69
-micromamba install -p $BUILD_PREFIX \
-    conda-forge/label/emscripten::flang==$PKG_VERSION \
-    -y --no-channel-priority
+if [ ! -f "$BUILD_PREFIX/bin/flang-new" ]; then
+    micromamba install -p $BUILD_PREFIX \
+        conda-forge/label/emscripten::flang==$PKG_VERSION \
+        llvm20==$PKG_VERSION \
+        -y --no-channel-priority
+fi
 
 export FC=flang-new
 export F77=flang-new
@@ -16,6 +19,7 @@ export FLANG=flang-new
 
 export FFLAGS="--target=wasm32-unknown-emscripten"
 export FPICFLAGS="-fPIC"
+export FCLIBS="-lFortranRuntime"
 
 rm $BUILD_PREFIX/bin/clang || true
 ln -s $BUILD_PREFIX/bin/clang-20 $BUILD_PREFIX/bin/clang # links to emsdk clang
